+++ tests/object_standards/test_sensemaker.py
@@ -1,10 +1,13 @@
+from datetime import datetime, timezone
 from unittest.mock import MagicMock, call, patch
 from uuid import uuid4
 
 import pytest
+from oms_sdk import DEFAULT_ACM
 from oms_sdk.generated.generated_graphql_client import (
     AttributeAttribute,
     NodesNodes,
+    ObjectStandardsObjectStandards,
     RelationshipRelationship,
 )
 
@@ -149,7 +152,9 @@ def test_process_data_attribute_passed_in(sensemaker):
     assert result == []
 
 
-def test_process_data_rel_passed_in(sensemaker):
+@patch("oms_sensemaking.object_standards.sensemaker.aac_client.get_acm_rollup")
+def test_process_data_rel_passed_in(mock_aac_client, sensemaker):
+    sensemaker.executed_at = datetime(2025, 5, 17, 14, 30, tzinfo=timezone.utc)  # set arbitrary time for sensemaker
     rel_of_node = MagicMock(spec=RelationshipRelationship)
     rel_of_node.startNodeId = str(uuid4())
     rel_of_node.endNodeId = str(uuid4())
@@ -157,10 +162,16 @@ def test_process_data_rel_passed_in(sensemaker):
     mock_nodes = MagicMock(spec=NodesNodes)
     mock_node = MagicMock()
     mock_node.classIri = "http://example.org/ClassIRI"
+    mock_node.acm = DEFAULT_ACM
     mock_node2 = MagicMock()
     mock_node2.classIri = "http://example.org/ClassIRI"
+    mock_node2.acm = DEFAULT_ACM
     mock_nodes.data = [mock_node, mock_node2]
     sensemaker.oms_crud_tool.get_nodes.return_value = mock_nodes
+
+    mock_obj_stds = MagicMock(spec=ObjectStandardsObjectStandards)
+    mock_obj_stds.data = []
+    sensemaker.oms_crud_tool.get_object_standards.return_value = mock_obj_stds
 
     required_attributes = ["iri1", "iri2"]
     required_relationships = ["relIri1", "relIri2"]
@@ -169,13 +180,18 @@ def test_process_data_rel_passed_in(sensemaker):
         return_value=RequiredIris(attribute_iris=required_attributes, relationship_iris=required_relationships)
     )
 
-    mock_attributes = [{"attribute": "value"}]
-    mock_relationships = [{"relationship": "value"}]
+    mock_attr = MagicMock(spec=AttributeAttribute)
+    mock_attr.acm = DEFAULT_ACM
+    mock_rel = MagicMock(spec=RelationshipRelationship)
+    mock_rel.acm = DEFAULT_ACM
+    mock_attributes = [mock_attr]
+    mock_relationships = [mock_rel]
 
     sensemaker.obj_standards_retriever.retrieve_data_for_grading.return_value = {
         "attributes": mock_attributes,
         "relationships": mock_relationships,
     }
+    mock_aac_client.return_value = DEFAULT_ACM
 
     mock_grade = MagicMock()
     sensemaker.obj_standards_rubric.grade = MagicMock(return_value=mock_grade)
@@ -286,3 +302,15 @@ def test_get_required_iris_no_parent_class():
     assert required_iris.attribute_iris == []
     assert required_iris.relationship_iris == []
     mock_ontology_service.get_node_ancestors_iris.assert_called_once_with(mock_node)
+
+
+def test_generate_summary_string(sensemaker):
+    float_score = 0.5
+    ratio_score = "1/2"
+    len_violations = 1
+    len_compliant_fields = 1
+    summary = sensemaker.generate_summary_string(float_score, ratio_score, len_violations, len_compliant_fields)
+    assert summary == (
+        "This object has an Object Standards score of 0.5 (1/2). "
+        "This object has 1 violation(s) and 1 compliant object(s)."
+    )